+++ setuptools_rust/build.py
@@ -1,6 +1,7 @@
 import glob
 import json
 import os
+import platform
 import shutil
 import sys
 import subprocess
@@ -95,6 +96,9 @@ class build_rust(RustCommand):
             target_triple = "i686-pc-windows-msvc"
         elif self.plat_name == "win-amd64":
             target_triple = "x86_64-pc-windows-msvc"
+        elif self.plat_name.startswith("macosx-") and platform.machine() == "x86_64":
+            # x86_64 or arm64 macOS targeting x86_64
+            target_triple = "x86_64-apple-darwin"
 
         if target_triple is not None:
             target_args = ["--target", target_triple]