+++ CHANGELOG.md
@@ -1,13 +1,20 @@
 # Changelog
 
 ## Unreleased
+### Packaging
+- Bump minimum Python version to Python 3.6.
 
+### Added
+- Support building x86-64 wheel on arm64 macOS machine. [#114](https://github.com/PyO3/setuptools-rust/pull/114)
+
+### Changed
 - Respect `PYO3_PYTHON` and `PYTHON_SYS_EXECUTABLE` environment variables if set. [#96](https://github.com/PyO3/setuptools-rust/pull/96)
 - Add runtime dependency on setuptools >= 46.1. [#102](https://github.com/PyO3/setuptools-rust/pull/102)
-- Append to, rather than replace, existing RUSTFLAGS when building. [#103](https://github.com/PyO3/setuptools-rust/pull/103)
+- Append to, rather than replace, existing `RUSTFLAGS` when building. [#103](https://github.com/PyO3/setuptools-rust/pull/103)
+
+### Fixed
 - Set executable bit on shared library. [#110](https://github.com/PyO3/setuptools-rust/pull/110)
-- Don't require optional wheel dependency. [#111](https://github.com/PyO3/setuptools-rust/pull/111)
-- Support building x86-64 wheel on arm64 macOS machine. [#114](https://github.com/PyO3/setuptools-rust/pull/114)
+- Don't require optional `wheel` dependency. [#111](https://github.com/PyO3/setuptools-rust/pull/111)
 
 ## 0.11.6 (2020-12-13)